I see this error with the app, reported by NextJS, please fix it. The error is reported as HTML but presented visually to the user).
A > before the line number in the error source usually indicates the line of interest:

> Internal Server Error
